[ypp-ops]: addresses https://github.com/Joystream/youtube-synch/issues/239
